ezoz/requirements.txt

asgiref==3.4.0
astroid==2.5.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
colorama==0.4.4
cryptography==36.0.1
defusedxml==0.7.1
Django==3.2.9
django-shopping-cart==0.1
djangorestframework==3.13.1
environ==1.0
idna==3.3
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
oauthlib==3.1.1
Pillow==8.4.0
pycparser==2.21
PyJWT==2.3.0
pylint==2.7.2
pymongo==3.11.4
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-core==4.1.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.7
wrapt==1.12.1